requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.6
django-extensions==3.1.3
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==8.12.1
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
python-http-client==3.3.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
sendgrid==6.8.0
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
